refactor(client): Replace ButtonProps
