+++ client/src/types/Button.ts
@@ -1,0 +1,5 @@
+export type ButtonProps = {
+  size?: "sm" | "md" | "lg";
+  variant?: "primary" | "danger";
+  loading?: boolean;
+};
